# Conflicts:
#	stroom-app/pom.xml
#	stroom-core-client-widget/pom.xml.versionsBackup
#	stroom-core-client/src/main/java/stroom/folder/client/FolderPlugin.java
#	stroom-core-server/pom.xml.versionsBackup
#	stroom-core-server/src/main/java/stroom/entity/server/ServiceCacheMethodInterceptor.java
#	stroom-dashboard-client/pom.xml.versionsBackup
#	stroom-integrationtest/src/test/java/stroom/feed/server/TestFeedServiceImpl.java
#	stroom-integrationtest/src/test/java/stroom/security/server/TestDocumentPermissionsServiceImpl.java
#	stroom-query-api/pom.xml
#	stroom-security-server/src/main/java/stroom/security/server/DocumentPermissionServiceImpl.java
#	stroom-security-server/src/main/java/stroom/security/server/SecurityContextImpl.java
#	stroom-security-shared/src/main/java/stroom/security/shared/UserRef.java
#	stroom-statistics-server/src/main/java/stroom/statistics/server/common/StatisticsDataSourceCacheImpl.java
#	stroom-statistics-shared/pom.xml.versionsBackup

+++ stroom-index-server/src/main/java/stroom/index/server/IndexShardWriterImpl.java
@@ -32,12 +32,12 @@ import org.apache.lucene.util.Version;
 import org.joda.time.DateTime;
 import stroom.index.server.analyzer.AnalyzerFactory;
 import stroom.index.shared.Index;
+import stroom.index.shared.IndexField;
+import stroom.index.shared.IndexField.AnalyzerType;
+import stroom.index.shared.IndexFields;
 import stroom.index.shared.IndexShard;
 import stroom.index.shared.IndexShard.IndexShardStatus;
 import stroom.index.shared.IndexShardService;
-import stroom.query.shared.IndexField;
-import stroom.query.shared.IndexField.AnalyzerType;
-import stroom.query.shared.IndexFields;
 import stroom.streamstore.server.fs.FileSystemUtil;
 import stroom.util.logging.LoggerPrintStream;
 import stroom.util.logging.StroomLogger;
@@ -319,8 +319,10 @@ public class IndexShardWriterImpl implements IndexShardWriter {
                     // is unlocked.
                     try {
                         final Path lockFile = dir.resolve(IndexWriter.WRITE_LOCK_NAME);
-                        Files.delete(lockFile);
-                    } catch (final InvalidPathException e) {
+                        if (Files.isRegularFile(lockFile)) {
+                            Files.delete(lockFile);
+                        }
+                    } catch (final IOException e) {
                         // There is no lock file so ignore.
                     }
 
@@ -721,11 +723,15 @@ public class IndexShardWriterImpl implements IndexShardWriter {
     private synchronized void refreshEntity() {
         try {
             // Update the size of the index.
-            if (dir != null) {
+            if (dir != null && Files.isDirectory(dir)) {
                 final AtomicLong totalSize = new AtomicLong();
-                Files.list(dir).forEach(file -> {
-                    totalSize.addAndGet(file.toFile().length());
-                });
+                try {
+                    Files.list(dir).forEach(file -> {
+                        totalSize.addAndGet(file.toFile().length());
+                    });
+                } catch (final IOException e) {
+                    LOGGER.error(e.getMessage());
+                }
                 indexShard.setFileSize(totalSize.get());
             }
